+++ lib/src/components/components.dart
@@ -4,5 +4,7 @@ export 'button/octo_button.dart';
 export 'flash/octo_flash.dart';
 export 'icon_button/octo_icon_button.dart';
 export 'label/octo_label.dart';
+export 'menu/octo_menu.dart';
+export 'menu/octo_menu_controller.dart';
 export 'text_field/octo_text_field.dart';
 export 'tooltip/octo_tooltip.dart';